requirements.txt
fastapi
openai
requests
python-dotenv
uvicorn
pytest
requests-mock
